Changing int i to long int i

+++ Lab-12/dining-philosopher-problem.c
@@ -47,11 +47,11 @@ void check_neighbours(int num){
 
 void *philosopher(void *arg){
     while(1){
-        int num = (int)arg; // Converting argument to integer
+        long int num = (long int)arg; // Converting argument to integer
         sleep(1);  // Thinking for some time
         
         state[num] = HUNGRY;    // Changing state to HUNGRY
-        printf("Philosopher-%d is hungry\n", num+1);
+        printf("Philosopher-%ld is hungry\n", num+1);
         
         sem_wait(&mutex);   // Acquiring mutex to pickup fork
         check_neighbours(num);  // Checking if neighbours are not eating and wake him if not
@@ -62,8 +62,8 @@ void *philosopher(void *arg){
 
         sem_wait(&mutex);   // Acquiring mutex to putdown fork
         state[num] = THINKING;  // Changing state to THINKING
-        printf("Philosopher-%d is putting down the forks\n", num+1);
-        printf("Philosopher-%d is thinking\n", num+1);
+        printf("Philosopher-%ld is putting down the forks\n", num+1);
+        printf("Philosopher-%ld is thinking\n", num+1);
         check_neighbours((num+N-1)%N);  // Checking if left neighbour can eat
         check_neighbours((num+1)%N);    // Checking if right neighbour can eat
         sem_post(&mutex);   // Releasing mutex to so that other philosophers can putdown fork
@@ -85,10 +85,10 @@ int main(){
 
     // Step-3: Creating a thread for each philosopher
     pthread_t philosophers[N];
-    for(int i=0; i<N; i++){
+    for(long int i=0; i<N; i++){
         state[i] = THINKING;    // Initializing state of each philosopher to THINKING
         pthread_create(&philosophers[i], NULL, philosopher, (void *)i);
-        printf("Philosopher-%d is thinking\n", i+1);
+        printf("Philosopher-%ld is thinking\n", i+1);
     }
 
     // Step-4: Joining all the threads
